MODORDERS-121 Create Instance Record in inventory when Order status is changed to 'Open'
- 'update Order' business logic updated to create Inventory records when order status is changed to Open
- 'create Order' business logic updated to handle Open orders via intermediate Pending state
