+++ src/main/java/org/folio/rest/impl/PostOrderLineHelper.java
@@ -1,25 +1,15 @@
 package org.folio.rest.impl;
 
-import static java.util.Collections.singletonList;
 import static java.util.concurrent.CompletableFuture.completedFuture;
-import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toMap;
-import static org.folio.orders.utils.HelperUtils.handleGetRequest;
 import static org.folio.orders.utils.HelperUtils.operateOnSubObj;
 import static org.folio.orders.utils.SubObjects.*;
-import static org.folio.rest.tools.client.Response.isSuccess;
 
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
-import java.util.concurrent.CompletionStage;
 
 import javax.ws.rs.core.Response;
 
-import org.folio.orders.rest.exceptions.HttpException;
 import org.folio.orders.utils.HelperUtils;
 import org.folio.rest.jaxrs.model.*;
 import org.folio.rest.jaxrs.model.Error;
@@ -31,35 +21,18 @@ import io.vertx.core.Context;
 import io.vertx.core.Handler;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpMethod;
-import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import me.escoffier.vertx.completablefuture.VertxCompletableFuture;
 
 public class PostOrderLineHelper extends AbstractHelper {
 
-  private static final String DEFAULT_INSTANCE_TYPE_CODE = "zzz";
-  private static final String DEFAULT_STATUS_CODE = "temp";
-  private static final String LOCATION_HEADER = "Location";
+
 
   PostOrderLineHelper(HttpClientInterface httpClient, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context ctx, String lang) {
     super(httpClient, okapiHeaders, asyncResultHandler, ctx, lang);
   }
 
-
-  public CompletableFuture<PoLine> createPoLine(PoLine compPOL, boolean updateInventory) {
-
-    if (updateInventory) {
-      return getProductTypesMap(compPOL)
-        .thenCompose(productTypesMap -> getInstanceRecord(compPOL, productTypesMap))
-        .thenCompose(instanceId -> createPoLine(compPOL, instanceId));
-    }
-    return createPoLine(compPOL, null);
-  }
-
-  private CompletableFuture<PoLine> createPoLine(PoLine compPOL, String instanceId) {
-    if (instanceId != null) {
-      compPOL.setInstanceId(instanceId);
-    }
+  public CompletableFuture<PoLine> createPoLine(PoLine compPOL) {
     JsonObject line = JsonObject.mapFrom(compPOL);
     List<CompletableFuture<Void>> subObjFuts = new ArrayList<>();
 
@@ -95,161 +68,7 @@ public class PostOrderLineHelper extends AbstractHelper {
       });
   }
 
-  /**
-   * Retrieves product type details associated with given PO line
-   * and builds 'product type name' -> 'product type id' map.
-   *
-   * @param compPOL the PO line to retrieve product type details for
-   * @return product types map
-   */
-  private CompletableFuture<Map<String, String>> getProductTypesMap(PoLine compPOL) {
-    // do not fail if no productId is provided, should be enforced on schema level if it's required
-    if (compPOL.getDetails() == null || compPOL.getDetails().getProductIds().isEmpty()) {
-      return completedFuture(Collections.emptyMap());
-    }
 
-    String endpoint = compPOL.getDetails().getProductIds().stream()
-      .map(productId -> String.format("name==%s", productId.getProductIdType().toString()))
-      .collect(joining(" or ", "/identifier-types?query=", ""));
-
-    return handleGetRequest(endpoint, httpClient, ctx, okapiHeaders, logger)
-      .thenApply(productTypes -> {
-        if (productTypes.getJsonArray("identifierTypes").size() != compPOL.getDetails().getProductIds().size()) {
-          throw new CompletionException(new HttpException(422,
-            "Invalid product type(s) is specified for the PO line with id " + compPOL.getId()));
-        }
-        return productTypes;
-      })
-      .thenApply(productTypes -> productTypes.getJsonArray("identifierTypes").stream()
-        .collect(toMap(jsonObj -> ((JsonObject) jsonObj).getString("name"),
-          jsonObj -> ((JsonObject) jsonObj).getString("id"),
-          (k1, k2) -> k1)));
-  }
-
-  /**
-   * Returns Id of the Instance Record corresponding to given PO line.
-   * Instance record is either retrieved from Inventory or a new one is created if no corresponding Record exists.
-   *
-   * @param compPOL PO line to retrieve Instance Record Id for
-   * @param productTypesMap product types Map used to build Inventory query
-   * @return future with Instance Id
-   */
-  private CompletionStage<String> getInstanceRecord(PoLine compPOL, Map<String, String> productTypesMap) {
-    // proceed with new Instance Record creation if no productId is provided
-    if (compPOL.getDetails() == null || compPOL.getDetails().getProductIds().isEmpty()) {
-      return createInstanceRecord(compPOL, productTypesMap);
-    }
-
-    String query = compPOL.getDetails().getProductIds().stream()
-      .map(productId -> buildProductIdQuery(productId, productTypesMap))
-      .collect(joining(" or "));
-
-    // query contains special characters so must be encoded before submitting
-    String endpoint = null;
-    try {
-      endpoint = "/inventory/instances?query=" + URLEncoder.encode(query, "UTF-8");
-    } catch (UnsupportedEncodingException e) {
-      logger.error(String.format("Error during query encoding: %s", e.getMessage()));
-      throw new CompletionException(e);
-    }
-
-    return handleGetRequest(endpoint, httpClient, ctx, okapiHeaders, logger)
-      .thenCompose(instances -> {
-        if (instances.getJsonArray("instances").size() > 0) {
-          return completedFuture(instances.getJsonArray("instances").getJsonObject(0).getString("id"));
-        }
-        return createInstanceRecord(compPOL, productTypesMap);
-      });
-  }
-
-  /**
-   * Creates Instance Record in Inventory and returns its Id.
-   *
-   * @param compPOL PO line to create Instance Record for
-   * @param productTypesMap product types Map used to build Instance Record json object
-   * @return id of newly created Instance Record
-   */
-  private CompletableFuture<String> createInstanceRecord(PoLine compPOL, Map<String, String> productTypesMap) {
-    JsonObject lookupObj = new JsonObject();
-    CompletableFuture<Void> instanceTypeFuture = getInstanceType(DEFAULT_INSTANCE_TYPE_CODE)
-      .thenAccept(lookupObj::mergeIn);
-    CompletableFuture<Void> statusFuture = getStatus(DEFAULT_STATUS_CODE)
-      .thenAccept(lookupObj::mergeIn);
-
-    return VertxCompletableFuture.allOf(ctx, instanceTypeFuture, statusFuture)
-      .thenApply(v -> buildInstanceRecordJsonObject(compPOL, productTypesMap, lookupObj))
-      .thenCompose(instanceRecJson -> {
-        try {
-          return httpClient.request(HttpMethod.POST, instanceRecJson.toBuffer(), "/inventory/instances", okapiHeaders)
-            .thenApply(response -> {
-              logger.debug("Validating received response");
-              if (!isSuccess(response.getCode())) {
-                throw new CompletionException(
-                  new HttpException(response.getCode(), response.getError().getString("errorMessage")));
-              }
-              return response;
-            })
-            .thenApply(response -> {
-              String location = response.getHeaders().get(LOCATION_HEADER);
-              return location.substring(location.lastIndexOf('/')+1);
-            });
-        } catch (Exception e) {
-          throw new CompletionException(e);
-        }
-      });
-  }
-
-  private CompletableFuture<JsonObject> getInstanceType(String typeName) {
-    String endpoint = String.format("/instance-types?query=code==%s", typeName);
-    return handleGetRequest(endpoint, httpClient, ctx, okapiHeaders, logger);
-  }
-
-  private CompletableFuture<JsonObject> getStatus(String statusCode) {
-    String endpoint = String.format("/instance-statuses?query=code==%s", statusCode);
-    return handleGetRequest(endpoint, httpClient, ctx, okapiHeaders, logger);
-  }
-
-  private String buildProductIdQuery(ProductId productId, Map<String, String> productTypes) {
-    return String.format("(identifiers adj \"\\\"identifierTypeId\\\": \\\"%s\\\"\" " +
-        "and identifiers adj \"\\\"value\\\": \\\"%s\\\"\")",
-      productTypes.get(productId.getProductIdType().toString()),
-      productId.getProductId());
-  }
-
-  private JsonObject buildInstanceRecordJsonObject(PoLine compPOL, Map<String, String> productTypes, JsonObject lookupObj) {
-    JsonObject instance = new JsonObject();
-    if (compPOL.getSource() != null) {
-      instance.put("source", compPOL.getSource().getCode());
-    }
-    if (compPOL.getTitle() != null) {
-      instance.put("title", compPOL.getTitle());
-    }
-    if (compPOL.getEdition() != null) {
-      instance.put("editions", new JsonArray(singletonList(compPOL.getEdition())));
-    }
-    instance.put("statusId", lookupObj.getJsonArray("instanceStatuses").getJsonObject(0).getString("id"));
-    instance.put("instanceTypeId", lookupObj.getJsonArray("instanceTypes").getJsonObject(0).getString("id"));
-
-    if (compPOL.getPublisher() != null || compPOL.getPublicationDate() != null) {
-      JsonObject publication = new JsonObject();
-      publication.put("publisher", compPOL.getPublisher());
-      publication.put("dateOfPublication", compPOL.getPublicationDate());
-      instance.put("publication", new JsonArray(Arrays.asList(publication)));
-    }
-
-    if (compPOL.getDetails() != null && compPOL.getDetails().getProductIds() != null) {
-      List<JsonObject> identifiers = compPOL.getDetails().getProductIds().stream()
-        .map(pId -> {
-          JsonObject identifier = new JsonObject();
-          identifier.put("identifierTypeId", productTypes.get(pId.getProductIdType().toString()));
-          identifier.put("value", pId.getProductId());
-          return identifier;
-        })
-        .collect(toList());
-      instance.put("identifiers", new JsonArray(identifiers));
-    }
-    return instance;
-  }
 
   private CompletableFuture<Void> createAdjustment(PoLine compPOL, JsonObject line, Adjustment adjustment) {
     return createSubObjIfPresent(line, adjustment, ADJUSTMENT, resourcesPath(ADJUSTMENT))
