+++ src/test/resources/mockdata/compositeOrders/1ab7ef6a-d1d4-4a4f-90a2-882aed18af14.json
@@ -20,7 +20,7 @@
   "total_estimated_price": 127.94,
   "total_items": 3,
   "vendor": "50fb88ea-cdf1-11e8-a8d5-f2801f1b9fd1",
-  "workflow_status": "Open",
+  "workflow_status": "Pending",
   "metadata": {
     "createdDate": "2018-06-29T00:00:00.000",
     "createdByUserId": "28d0f99c-d137-11e8-a8d5-f2801f1b9fd1"
